+++ clearwsd-verbnet/src/main/java/io/github/clearwsd/verbnet/ThematicRole.java
@@ -16,6 +16,7 @@
 
 package io.github.clearwsd.verbnet;
 
+import io.github.clearwsd.verbnet.restrictions.Restrictions;
 import java.util.List;
 
 /**
