+++ clearwsd-verbnet/src/main/java/io/github/clearwsd/verbnet/VerbNetFrame.java
@@ -16,6 +16,8 @@
 
 package io.github.clearwsd.verbnet;
 
+import io.github.clearwsd.verbnet.semantics.SemanticPredicate;
+import io.github.clearwsd.verbnet.syntax.SyntacticPhrase;
 import java.util.List;
 
 /**
